+++ src/help/help.service.spec.ts
@@ -25,7 +25,7 @@ describe("HelpService", () => {
     const testingModule = await Test.createTestingModule({
       imports: [
         CacheModule.register({ isGlobal: true, store }),
-        TelegrafModule.forRoot({ launchOptions: false, token: process.env.TELEGRAM_TOKEN ?? "" }),
+        TelegrafModule.forRoot({ launchOptions: false, token: process.env.TG_TOKEN ?? "" }),
         PrismaModule,
         SettingsModule,
       ],