+++ src/messages/messages.service.spec.ts
@@ -14,7 +14,7 @@ describe("MessagesService", () => {
     const testingModule = await Test.createTestingModule({
       imports: [
         CacheModule.register({ isGlobal: true, store }),
-        TelegrafModule.forRoot({ launchOptions: false, token: process.env.TELEGRAM_TOKEN ?? "" }),
+        TelegrafModule.forRoot({ launchOptions: false, token: process.env.TG_TOKEN ?? "" }),
         PrismaModule,
       ],
       providers: [MessagesService],