+++ test/help.e2e-spec.ts
@@ -7,11 +7,7 @@ import type { App } from "supertest/types";
 import * as fixtures from "fixtures/help";
 import { AppModule } from "src/app.module";
 
-import {
-  TEST_TELEGRAM_API_BASE_URL,
-  TEST_TELEGRAM_WEBHOOK_BASE_URL,
-  TEST_TELEGRAM_WEBHOOK_PATH,
-} from "./utils/constants";
+import { TEST_TG_API_BASE_URL, TEST_TG_WEBHOOK_BASE_URL, TEST_TG_WEBHOOK_PATH } from "./utils/constants";
 import { server } from "./utils/server";
 
 describe("HelpModule (e2e)", () => {
@@ -28,14 +24,14 @@ describe("HelpModule (e2e)", () => {
   it("answers to /help command in a supergroup chat", async () => {
     let sendMessagePayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/sendMessage`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/sendMessage`, async (info) => {
         sendMessagePayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.supergroupHelpWebhook);
 
     expect(response.status).toBe(200);
@@ -46,7 +42,7 @@ describe("HelpModule (e2e)", () => {
     let sendMessagePayload1: unknown;
     let sendMessagePayload2: unknown;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/sendMessage`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/sendMessage`, async (info) => {
         const body = await info.request.json();
         sendMessagePayload2 = sendMessagePayload1 ? body : undefined;
         sendMessagePayload1 = sendMessagePayload1 ?? body;
@@ -54,8 +50,8 @@ describe("HelpModule (e2e)", () => {
       }),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.privateStartWebhook);
 
     expect(response.status).toBe(200);
@@ -66,14 +62,14 @@ describe("HelpModule (e2e)", () => {
   it("ignores /help command if it has payload", async () => {
     let sendMessagePayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/sendMessage`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/sendMessage`, async (info) => {
         sendMessagePayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.helpWithPayloadWebhook);
 
     expect(response.status).toBe(200);