+++ test/profanity-filter.e2e-spec.ts
@@ -12,9 +12,9 @@ import { AppModule } from "src/app.module";
 
 import {
   TEST_ASYNC_DELAY,
-  TEST_TELEGRAM_API_BASE_URL,
-  TEST_TELEGRAM_WEBHOOK_BASE_URL,
-  TEST_TELEGRAM_WEBHOOK_PATH,
+  TEST_TG_API_BASE_URL,
+  TEST_TG_WEBHOOK_BASE_URL,
+  TEST_TG_WEBHOOK_PATH,
 } from "./utils/constants";
 import { createDbSupergroupChat, createDbUser, prisma } from "./utils/database";
 import { server } from "./utils/server";
@@ -36,8 +36,8 @@ describe("ProfanityFilterModule (e2e)", () => {
     await prisma.profaneWord.create({
       data: { authorId: adminUser.id, editorId: adminUser.id, language: LanguageCode.EN, word: "bad" },
     });
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.channelMessageWebhook);
 
     expect(response.status).toBe(200);
@@ -47,8 +47,8 @@ describe("ProfanityFilterModule (e2e)", () => {
   it("handles an error if chat id is incorrect during settings rendering", async () => {
     const stderrWriteSpy = jest.spyOn(process.stderr, "write").mockImplementation(() => true);
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSettingsErrorWebhook);
     expect(response.status).toBe(200);
     expect(stderrWriteSpy).toHaveBeenCalledTimes(1);
@@ -57,8 +57,8 @@ describe("ProfanityFilterModule (e2e)", () => {
   it("handles an error if chat id is incorrect during settings saving", async () => {
     const stderrWriteSpy = jest.spyOn(process.stderr, "write").mockImplementation(() => true);
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSaveSettingsErrorWebhook);
     expect(response.status).toBe(200);
     expect(stderrWriteSpy).toHaveBeenCalledTimes(1);
@@ -68,14 +68,14 @@ describe("ProfanityFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let editMessageTextPayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({}, { status: 400 });
       }),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSettingsWebhook);
 
     expect(response.status).toBe(200);
@@ -87,14 +87,14 @@ describe("ProfanityFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let editMessageTextPayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSaveSettingsWebhook);
 
     const expectedEditMessageTextPayload = fixtures.cbSaveSettingsEditMessageTextPayload();
@@ -110,8 +110,8 @@ describe("ProfanityFilterModule (e2e)", () => {
       data: { authorId: adminUser.id, editorId: adminUser.id, language: LanguageCode.EN, word: "bad" },
     });
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.autoForwardChannelMessageWebhook);
 
     expect(response.status).toBe(200);
@@ -124,8 +124,8 @@ describe("ProfanityFilterModule (e2e)", () => {
       data: { authorId: adminUser.id, editorId: adminUser.id, language: LanguageCode.EN, word: "bad" },
     });
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.channelMessageWebhook);
 
     expect(response.status).toBe(200);
@@ -135,17 +135,15 @@ describe("ProfanityFilterModule (e2e)", () => {
   it("should not render settings if the user is not an admin", async () => {
     let editMessageTextPayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/getChatAdministrators`, () =>
-        HttpResponse.json({ ok: true, result: [] }),
-      ),
+      http.post(`${TEST_TG_API_BASE_URL}/getChatAdministrators`, () => HttpResponse.json({ ok: true, result: [] })),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSettingsWebhook);
 
     expect(response.status).toBe(200);
@@ -158,17 +156,15 @@ describe("ProfanityFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let editMessageTextPayload;
     server.use(
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TEST_TG_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
-      http.post(`${TEST_TELEGRAM_API_BASE_URL}/getChatAdministrators`, () =>
-        HttpResponse.json({ ok: true, result: [] }),
-      ),
+      http.post(`${TEST_TG_API_BASE_URL}/getChatAdministrators`, () => HttpResponse.json({ ok: true, result: [] })),
     );
 
-    const response = await request(TEST_TELEGRAM_WEBHOOK_BASE_URL)
-      .post(TEST_TELEGRAM_WEBHOOK_PATH)
+    const response = await request(TEST_TG_WEBHOOK_BASE_URL)
+      .post(TEST_TG_WEBHOOK_PATH)
       .send(fixtures.cbSaveSettingsWebhook);
 
     expect(response.status).toBe(200);